flask_start/templates/common/requirements.txt
path.py
Flask
Flask-Login
Flask-Principal
Flask-SQLAlchemy
Flask-WTF
Flask-Babel
